+++ extensions/music/package.json
@@ -290,9 +290,9 @@
       ]
     },
     {
-      "name": "now-playing-menu-bar",
+      "name": "currently-playing-menu-bar",
       "title": "Menu Bar Player",
-      "description": "See what's currently playing in your Menu Bar. Click for further actions, such as Play/Pause, Like/Dislike, Skip, Start Radio, Add to Playlist, Connect Device and more. Refreshes every 10 seconds.",
+      "description": "See what's currently playing in your Menu Bar. Click for further actions, such as Play/Pause. Refreshes every 10 seconds.",
       "mode": "menu-bar",
       "interval": "10s",
       "preferences": [
